# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ARTEMIS Phase 1",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/state": {
      "get": {
        "summary": "Get Session State",
        "description": "Return the authoritative session snapshot for full resync.\n\nCalled by the frontend when the WS event bus sends ``client.resync_required``\n(i.e. the requested replay window is no longer available).\n\nSee ``docs/api.md`` \u00a73 and \u00a74.",
        "operationId": "get_session_state_v1_sessions__session_id__state_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/models/select": {
      "post": {
        "summary": "Select Model",
        "operationId": "select_model_v1_models_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSelectionReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/approvals/{approval_id}": {
      "post": {
        "summary": "Respond Approval",
        "operationId": "respond_approval_v1_approvals__approval_id__post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalResponseReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/grants/{grant_id}": {
      "delete": {
        "summary": "Revoke Grant",
        "operationId": "revoke_grant_v1_grants__grant_id__delete",
        "parameters": [
          {
            "name": "grant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/settings/fs/roots": {
      "post": {
        "summary": "Add Fs Root",
        "operationId": "add_fs_root_v1_settings_fs_roots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRootReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Fs Root",
        "operationId": "remove_fs_root_v1_settings_fs_roots_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRootReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddRootReq": {
        "properties": {
          "root": {
            "type": "string",
            "title": "Root"
          },
          "confirm_risk": {
            "type": "boolean",
            "title": "Confirm Risk",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "root"
        ],
        "title": "AddRootReq"
      },
      "ApprovalResponseReq": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "once"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ApprovalResponseReq"
      },
      "AssistantStateData": {
        "properties": {
          "state": {
            "title": "State",
            "type": "string"
          },
          "intensity": {
            "title": "Intensity",
            "type": "number"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Progress"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Run Id"
          }
        },
        "required": [
          "state",
          "intensity"
        ],
        "title": "AssistantStateData",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelSelectionReq": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          }
        },
        "type": "object",
        "required": [
          "role",
          "model_id"
        ],
        "title": "ModelSelectionReq"
      },
      "RemoveRootReq": {
        "properties": {
          "root": {
            "type": "string",
            "title": "Root"
          }
        },
        "type": "object",
        "required": [
          "root"
        ],
        "title": "RemoveRootReq"
      },
      "SessionStateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "last_seq": {
            "type": "integer",
            "title": "Last Seq"
          },
          "assistant_state": {
            "$ref": "#/components/schemas/AssistantStateData"
          },
          "active_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run"
          },
          "pending_approvals": {
            "items": {},
            "type": "array",
            "title": "Pending Approvals"
          },
          "active_task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Task"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "last_seq",
          "assistant_state"
        ],
        "title": "SessionStateResponse",
        "description": "Authoritative session snapshot returned by GET /v1/sessions/{id}/state.\n\nUsed by the frontend to fully resync after a replay gap.\nSee ``docs/api.md`` \u00a73 and the ``client.resync_required`` WS event."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}